requirements.txt
Flask>=2.0
pytesseract
pillow
python-dotenv
rapidfuzz
pytest
